Flask后端/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
altgraph==0.17.2
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
blobfile==2.0.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
config==0.5.1
constantly==15.1.0
contourpy==1.0.6
creart==0.2.2
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
docopt==0.6.2
edge-tts==6.1.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fake-useragent==0.1.11
filelock==3.8.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
huggingface-hub==0.11.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.0.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jmespath==1.0.1
jupyter_core==5.1.0
keras==2.11.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==15.0.6.1
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
OpenAIAuth==0.0.5
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
parsel==1.7.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.13
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.36
Protego==0.2.1
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.17
PyDispatcher==2.0.6
pydocstyle==6.2.3
pyflakes==3.0.1
Pygments==2.13.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.9
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
pytoolconfig==1.2.5
pytz==2022.6
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
queuelib==1.6.2
redis==4.5.1
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
revChatGPT==0.0.34
rfc3986==1.5.0
rsa==4.9
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
some-package==0.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.4
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tiktoken==0.3.0
tldextract==3.4.0
tls-client==0.1.5
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.25.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.9
w3lib==2.1.1
wcwidth==0.2.5
Werkzeug==2.2.2
whatthepatch==1.0.4
wrapt==1.14.1
yapf==0.32.0
yarg==0.1.9
yarl==1.8.2
zipp==3.12.0
zope.interface==5.5.2
